# impressions-library-openapi.yaml
openapi: 3.0.3
info:
  title: L3 Miage projet BD
  description: |
    Every contract of our print app
  version: 1.0.0
servers:
  - url: "http://localhost:8080"
paths:
...
  /api/client/{clientId}/impressions/tirages:
    parameters:
      - name: clientId
        description: Client's id
        in: path
        required: true
        schema:
          type: number
          format: int64
    get:
      summary: Find all tirages
      operationId: find-all-tirages
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tirage"
        404:
          description: Client id was not found
    post:
      summary: Create a new impression of a Tirage for the client identified by the id in the path
      operationId: new-tirage-clients
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/NewTirage"
      responses:
        201:
          description: Created
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Tirage"
        400:
          description: In case of Tirage objet type in HTTP request could not be validated
        404:
          description: Client id was not found
    put:
      summary: Update a tirage (pas implémenté car pas dans le cahier des charges)
      operationId: update-tirage-client
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/Tirage"
      responses:
        200:
          description: Updated
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Tirage"
        400:
          description: error request HTTP
        404:
          description: Client could not be found
  /api/client/{clientId}/impressions/tirages/{tirId}/selections:
    parameters:
      - name: clientId
        description: Client's id
        in: path
        required: true
        schema:
          type: number
          format: int64
      - name: tirId
        description: tirage's id
        in: path
        required: true
        schema:
          type: number
          format: int64
    post:
      summary: Create a new selection tirage
      operationId: new-selection-tirage-clients
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/SelectionTirage"
      responses:
        201:
          description: Created
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Tirage"
        400:
          description: Tirage does not belong to the client or error HTTP request
        404:
          description: Client or tirage or photo id could not be found
...
components:
  schemas:
    NewCommande:
      description: A commande without id for creation
      type: object
      properties:
        dateDeCommande:
          type: string
          format: date
        prixTotal:
          type: number
          minimum: 0
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Impression"
        statut:
          type: string
          enum: [enattente, encours, termine]
      required:
        - dateDeCommande
        - articles
        - statut
      example:
        dateDeCommande: '2022-07-01'
        prixTotal: 20.0
        articles:
          - id: 2
        statut: encours
    Commande:
      description: A commande without id for creation
      allOf:
        - $ref: "#/components/schemas/NewCommande"
        - type: object
          properties:
            id:
              type: integer
              format: int64
          required:
            - id
          example:
            id: 1
    NewClient:
      description: A client without id for creation
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        adresseMail:
          type: string
        adressePostale:
          type: string
        motDePasse:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        impressions:
          type: array
          items:
            $ref: "#/components/schemas/Impression"
        commandes:
          type: array
          items:
            $ref: "#/components/schemas/Commande"
      required:
        - nom
        - prenom
        - adresseMail
        - adressePostale
        - motDePasse
      example:
        nom: Smith
        prenom: Bob
        adresseMail: bob.smith@wanadoo.fr
        adressePostale: 4 avenue lambda
        motDePasse: azerty123
        images:
          - id: 2
        impressions:
          - id: 3
        commandes:
          - id: 4
    Client:
      description: A client with an id for response
      allOf:
        - $ref: "#/components/schemas/NewClient"
        - type: object
          properties:
            id:
              type: integer
              format: int64
          required:
            - id
          example:
            id: 1
    NewImage:
      description: An image without id for creation
      properties:
        chemin:
          type: string
        info:
          type: string
        estPartage:
          type: boolean
        proprietaire:
          $ref: "#/components/schemas/Client"
      required:
        - chemin
        - estPartage
        - proprietaire
      default:
        estPartage: false
      example:
        chemin: nomsite.fr/path
        info: une image de type magnifique
        estPartage: false
        proprietaire:
          - id: 2
    Image:
      description: An image with an id for response
      allOf:
        - $ref: "#/components/schemas/NewImage"
        - type: object
          properties:
            id:
              type: integer
              format: int64
          required:
            - id
          example:
            id: 1
    NewPhoto:
      description: A photo without id for creation (on peut limiter le nombre de caractères des paramètres?)
      type: object
      properties:
        parametres:
          type: string
        imageSource:
          $ref: "#/components/schemas/Image"
      required:
        - imageSource
        - parametres
      example:
        parametres: grand;etoile
        imageSource:
          - id: 2
    Photo:
      description: An existing photo with an id
      allOf:
        - $ref: "#/components/schemas/NewPhoto"
        - type: object
          properties:
            id:
              type: integer
              format: int64
          required:
            - id
          example:
            id: 1
...
    NewSelectionTirage:
      description: A selection tirage without id for creation
      type: object
      properties:
        quantite:
          type: integer
          format: int64
        photo:
          $ref: "#/components/schemas/Photo"
      required:
        - quantite
        - photo
      example:
        quantite: 3
        photo:
          - id: 2
    SelectionTirage:
      description: A selection tirage with an id for update
      allOf:
        - $ref: "#/components/schemas/NewSelectionTirage"
        - type: object
          properties:
            id:
              type: integer
              format: int64
          required:
            - id
          example:
            id: 1
    BaseNewImpression:
      description: The 2 elements in common of any new type of impression creation. Pour éviter la redondance dans la ré-écriture
      type: object
      properties:
        format:
          type: string
          enum: [petit, moyen, grand]
        qualite:
          type: string
          enum: [faible, moyenne, bonne]
      required:
        - format
        - qualite
      example:
        format: moyen
        qualite: moyenne
    Impression:
      description: Pour éviter la redondance. Les 4 type d'impressions y font référence. For response only
      allOf:
        - $ref: "#/components/schemas/BaseNewImpression"
        - type: object
          properties:
            id:
              type: integer
              format: int64
            statut:
              type: string
              enum: [enattente, encours, termine]
            prix:
              type: number
              minimum: 0
            format:
              type: string
            qualite:
              type: string
            consommateur:
              $ref: "#/components/schemas/Client"
          required:
            - id
            - statut
            - prix
            - consommateur
          example:
            id: 1
            statut: encours
            prix: 20.0
            format: moyen
            qualite: moyenne
            consommateur:
              - id: 2
...
    NewTirage:
      description: A Tirage with no id (for creation)
      allOf:
        - $ref: "#/components/schemas/BaseNewImpression"
    Tirage:
      description: A Tirage with an id for response
      allOf:
        - $ref: "#/components/schemas/Impression"
        - type: object
          properties:
            selections:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/SelectionTirage"
          required:
            - selections
          example:
            selections:
              - id: 3
...